Refactoring pythonLib
