+++ .flake8
@@ -4,3 +4,4 @@ exclude = docs, .git, .idea, __pycache__
 ignore = E203 W291 W293 E743 E265 W605 W503
 max-line-length = 120
 statistics = True
+per-file-ignores = __init__.py: F401
